Functional Update
Added lots of files and function to lib directory useful for the initialization process. Most are files proveded with the kinarm itself. 

Renamed PlotTargetAnimation to Replay. The function was also extensively restructured with a more sensible code. Next version will use a timer to run the animation. 

Added transient properties to Session and Subject and restructured the save functions.

Bugfixed plotForceDistribution and getLateralDeviation